Solucionando error Integer a Long e implementando el "salvavidas" de poder usar en la nube y local

+++ src/main/java/com/climatizacion/sistema_clima/dto/CitaRequestDTO.java
@@ -10,7 +10,7 @@ public class CitaRequestDTO {
     @NotNull(message = "El ID del cliente es obligatorio")
     private Long idCliente;
 
-    private Integer idPedido;
+    private Long idPedido;
 
     @NotNull(message = "El ID del técnico es obligatorio")
     private Long idTecnico;
