+++ src/main/java/com/climatizacion/sistema_clima/entities/CitaEntity.java
@@ -16,7 +16,7 @@ public class CitaEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_cita")
-    private Integer idCita;
+    private Long idCita;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_cliente", nullable = false)
