+++ src/main/java/com/climatizacion/sistema_clima/repository/SolicitudServicioRepository.java
@@ -4,7 +4,7 @@ import com.climatizacion.sistema_clima.entities.SolicitudServicioEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface SolicitudServicioRepository extends JpaRepository<SolicitudServicioEntity, Integer> {
+public interface SolicitudServicioRepository extends JpaRepository<SolicitudServicioEntity, Long> {
     List<SolicitudServicioEntity> findByEstadoOrderByFechaCreacionAsc(String estado);
     List<SolicitudServicioEntity> findByCliente_IdCliente(Long idCliente);
     long countByEstado(String estado);
